pyproject.toml
[tool.poetry]
name = "qna"
version = "0.1.0"
description = ""
authors = ["wjlee <wjlee.ling@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10, <3.12"
python-dotenv = "^1.0.1"
langchain = "^0.3.4"
langchain-openai = "^0.2.3"
langchain-pinecone = "^0.2.0"
pinecone-text = "^0.9.0"
kiwipiepy = "^0.19.1"
streamlit = "^1.39.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
